Changes according to code review
