+++ advanced-sample/src/main/kotlin/me/aartikov/replica/advanced_sample/features/fruits/data/favourite/FavouriteFruitsRepositoryImpl.kt
@@ -1,16 +1,19 @@
 package me.aartikov.replica.advanced_sample.features.fruits.data.favourite
 
-import me.aartikov.replica.advanced_sample.features.fruits.data.SetFruitIsFavouriteAction
+import kotlinx.coroutines.delay
+import me.aartikov.replica.advanced_sample.features.fruits.data.SetFruitFavouriteAction
 import me.aartikov.replica.advanced_sample.features.fruits.data.api.FruitApi
 import me.aartikov.replica.advanced_sample.features.fruits.data.dto.toDomain
 import me.aartikov.replica.advanced_sample.features.fruits.domain.FruitId
 import me.aartikov.replica.advanced_sample.features.fruits.domain.withUpdatedIsFavourite
 import me.aartikov.replica.client.ReplicaClient
+import me.aartikov.replica.client.sendActionWithOptimisticUpdate
 import me.aartikov.replica.common.OptimisticUpdate
 import me.aartikov.replica.single.ReplicaSettings
 import me.aartikov.replica.single.behaviour.ReplicaBehaviour
 import me.aartikov.replica.single.behaviour.standard.doOnAction
 import me.aartikov.replica.single.behaviour.standard.provideOptimisticUpdate
+import kotlin.time.Duration
 import kotlin.time.Duration.Companion.seconds
 
 class FavouriteFruitsRepositoryImpl(
@@ -24,14 +27,14 @@ class FavouriteFruitsRepositoryImpl(
             api.getFavouriteFruits().map { it.toDomain() }
         },
         behaviours = listOf(
-            ReplicaBehaviour.provideOptimisticUpdate { action: SetFruitIsFavouriteAction ->
+            ReplicaBehaviour.provideOptimisticUpdate { action: SetFruitFavouriteAction ->
                 OptimisticUpdate { fruits ->
                     fruits.withUpdatedIsFavourite(action.fruitId, action.isFavourite)
                 }
             },
-            ReplicaBehaviour.doOnAction { action: SetFruitIsFavouriteAction ->
+            ReplicaBehaviour.doOnAction { action: SetFruitFavouriteAction ->
                 when (action.isFavourite) {
-                    true -> refresh()
+                    true -> invalidate()
                     false -> mutateData { favourites ->
                         favourites.filter { it.id != action.fruitId }
                     }
@@ -40,13 +43,20 @@ class FavouriteFruitsRepositoryImpl(
         )
     )
 
-    override suspend fun removeFromFavourite(fruitId: FruitId) {
-        api.dislikeFruit(fruitId.value)
-        replicaClient.sendAction(
-            SetFruitIsFavouriteAction(
-                fruitId = fruitId,
-                isFavourite = false
-            )
-        )
+    override suspend fun setFruitFavourite(
+        fruitId: FruitId,
+        isFavourite: Boolean,
+        debounceDelay: Duration
+    ) {
+        replicaClient.sendActionWithOptimisticUpdate(
+            action = SetFruitFavouriteAction(fruitId, isFavourite)
+        ) {
+            delay(debounceDelay)
+            if (isFavourite) {
+                api.likeFruit(fruitId.value)
+            } else {
+                api.dislikeFruit(fruitId.value)
+            }
+        }
     }
 }
