+++ advanced-sample/src/main/kotlin/me/aartikov/replica/advanced_sample/features/fruits/ui/FruitsUi.kt
@@ -15,6 +15,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
 import com.arkivanov.decompose.extensions.compose.jetpack.stack.Children
 import me.aartikov.replica.advanced_sample.core.theme.AppTheme
 import me.aartikov.replica.advanced_sample.features.fruits.ui.all.FruitsAllUi
@@ -71,6 +72,7 @@ private fun FruitsTopBar(
                 textAlign = TextAlign.Center,
                 modifier = Modifier
                     .weight(1f)
+                    .padding(8.dp)
                     .clickable { onTabClick(tab) }
             )
         }
